requirements.txt
discord.py
youtube_dl
bs4
requests